6_ft_transcendence/backend/user/requirements.txt
django==5.1.5
django-filter
djangorestframework
requests
pyjwt
python-dotenv
pillow
psycopg2
# djangorestframework-simplejwt
# djangorestframework-queryfields
# django-cors-headers
# django-otp
# django-two-factor-auth
# qrcode